Lots of Stuff Changed
First off, I changed the TeleOp programs to what the second driver
wants, easy. Autonomous has also been added. The old OpModes have been
deleted. For now, we are working in RedAutoRamp, it won’t go to the
ramp, as we have decided to go for dumping people since the naming of
the class. In addition to this, I’ve added BaseAuto, an abstract class
extending LinearOpMode, and is what all of our individual autonomous
programs will extend. In BaseAuto, we have movement methods for moving
forward and backward and for turning in place. Each subclass of
BaseAuto will have to override recuriseveAuto which will be where we
will program all the movements. Take note of how little there is in
runOpMode in the subclass of BaseAuto. With our extension, the
declaration of all variables is done in the superclass, meaning that we
won’t have to worry about instanciating them with each new OpMode. When
making a new subclass of BaseAuto, be sure to call super.runOpMode();
as this will construct all of our variables… which we kind of need.
It doesn’t work perfectly yet, but it should be good by thursday